backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.5
Pillow==10.1.0
jdatetime==4.1.1
python-decouple==3.8

# Production dependencies
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
redis==5.0.1
django-redis==5.4.0
